Moved most of functions to Repository
