+++ src/core/Application.kt
@@ -1,7 +1,5 @@
 package core
 
-import core.Server
-
 fun main() {
 
     val server = Server()
